+++ programs/ekr_client.c
@@ -142,7 +142,7 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 			       rcv.rcv_sid,
 			       rcv.rcv_ssn,
 			       rcv.rcv_tsn,
-			       (uint32_t) ntohl(rcv.rcv_ppid),
+			       ntohl(rcv.rcv_ppid),
 			       rcv.rcv_context);
 		}
 		free(data);
